requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.34
pyyaml>=6.0
pydantic>=2.0
pydantic-settings>=2.0
python-dotenv>=1.0
httpx>=0.27
